fix(project): inspect returns main (no CLI on require) + shared projector; reject extra roots
Two integration fixes on Slice 2.

1. Tool-module convention + reusable projector. inspect.lua executed main() during
   require, contrary to Hull's convention (the tool dispatcher does `return require(mod)`
   and invokes the returned main only when it is the ENTRY command, so a dependency
   require never runs the CLI -- as hull.build / hull.init do with `return main`). This
   mattered immediately for Slice 3: discovery.json needs the IDENTICAL public
   projection, and requiring inspect.lua would have run the CLI. Fixed by extracting the
   wire schema into a SINGLE side-effect-free module hull.project.projection
   (M.project(disc) -> plain table; drops per-decl handle / _by_source / _handles / the
   by_id decl map). inspect.lua now requires it, builds main, and RETURNS main (never
   calls it). Both standalone inspection and (Slice 3) dev publication call the one
   projector -- no duplicated wire-schema logic.

2. Positional-count validation. `hull agent inspect a b` silently used the last root; it
   is now an explicit usage error (exit 2). At most one positional app_dir is accepted.

Also updated the design record's top-level status (Slice 1 merged #356, Slice 2
implemented #357) and documented the shared projection module + tool-module convention in
the Slice 2 record.

Tests: unit -- projection drops all generation-internal state while keeping the public
wire fields (test_project 82 -> 89); e2e -- two positional roots -> exit 2, single
positional still exit 0, unknown flag -> exit 2 (e2e 25 -> 28). 76/76 unit; luacheck
clean.
